requirements.txt
Django==1.7.1
-e git+git@github.com:jonykalavera/python-shotgunorm.git@3b0a2b433030815631588ff709c8ffd3e9660476#egg=ShotgunORM-master
django-tastypie==0.12.1
python-dateutil==2.2
python-mimeparse==0.1.4
-e git+git@github.com:shotgunsoftware/python-api.git@ebb8f0bb6e970043e22b8240df3aee2834cfdda2#egg=shotgun_api3-master
six==1.8.0
